Fix WizardModal's UI/UX issues
