+++ src/fullPageBookingForm/Components/Common/WizardModal/WizardModal.tsx
@@ -1,7 +1,12 @@
 /** @jsx h */
 import { h, FunctionComponent } from "preact";
-import { useEffect } from "preact/hooks";
-import { WizardModalProvider } from "./WizardModalProvider";
+import { useEffect, useRef } from "preact/hooks";
+import { CloseIcon } from "../Icon/CloseIcon";
+import {
+  WizardModalProvider,
+  useWizardModalState,
+  useWizardModalAction,
+} from "./WizardModalProvider";
 import { WizardModalPage } from "./WizardModalPage";
 import "./WizardModal.scss";
 
@@ -11,6 +16,48 @@ export type WizardModalProps = {
   onClose: () => void;
 };
 
+const Modal: FunctionComponent<Pick<WizardModalProps, "open">> = ({
+  children,
+  open,
+}) => {
+  const modalRef = useRef<HTMLDivElement>();
+  const { currentPage } = useWizardModalState();
+  const { close } = useWizardModalAction();
+  const wizardModalClassNames = ["wizard-modal"];
+
+  useEffect(() => {
+    modalRef.current?.scrollTo?.({
+      top: 0,
+    });
+  }, [currentPage]);
+
+  if (open) {
+    wizardModalClassNames.push("wizard-modal--open");
+  }
+
+  const handleCloseButtonClick = () => close();
+
+  return (
+    <div
+      role="dialog"
+      className={wizardModalClassNames.join(" ")}
+      ref={modalRef}
+    >
+      <div className="wizard-modal__title-bar" />
+      <div className="wizard-modal__body">
+        <div
+          data-testid="wizard-modal-close-button"
+          className="wizard-modal__close-button"
+          onClick={handleCloseButtonClick}
+        >
+          <CloseIcon color="#666" width={32} height={32} strokeSize={2} />
+        </div>
+        {children}
+      </div>
+    </div>
+  );
+};
+
 export const WizardModal: FunctionComponent<WizardModalProps> & {
   Page: typeof WizardModalPage;
 } = ({ open, initialPage, children, onClose }) => {
@@ -18,21 +65,13 @@ export const WizardModal: FunctionComponent<WizardModalProps> & {
     document.body.style.overflow = open ? "hidden" : "auto";
   }, [open]);
 
-  const wizardModalClassNames = ["wizard-modal"];
-
-  if (open) {
-    wizardModalClassNames.push("wizard-modal--open");
-  }
-
   return (
     <WizardModalProvider
       open={open}
       initialPage={initialPage}
       onClose={onClose}
     >
-      <div role="dialog" className={wizardModalClassNames.join(" ")}>
-        {children}
-      </div>
+      <Modal open={open}>{children}</Modal>
     </WizardModalProvider>
   );
 };
